[51] User access to endpoints
